p2p: consolidate protocol flows under v10
