+++ protocol/flows/src/lib.rs
@@ -5,5 +5,3 @@ pub mod ibd;
 pub mod service;
 pub mod user_agent_rule;
 pub mod v10;
-pub mod v7;
-pub mod v8;
