--- protocol/flows/src/v8/mod.rs
@@ -1,165 +1,0 @@
-use crate::v7::{
-    address::{ReceiveAddressesFlow, SendAddressesFlow},
-    blockrelay::{flow::HandleRelayInvsFlow, handle_requests::HandleRelayBlockRequests},
-    ping::{ReceivePingsFlow, SendPingsFlow},
-    request_antipast::HandleAntipastRequests,
-    request_block_locator::RequestBlockLocatorFlow,
-    request_headers::RequestHeadersFlow,
-    request_ibd_blocks::HandleIbdBlockRequests,
-    request_ibd_chain_block_locator::RequestIbdChainBlockLocatorFlow,
-    request_pp_proof::RequestPruningPointProofFlow,
-    request_pruning_point_and_anticone::PruningPointAndItsAnticoneRequestsFlow,
-    request_pruning_point_utxo_set::RequestPruningPointUtxoSetFlow,
-    txrelay::flow::{RelayTransactionsFlow, RequestTransactionsFlow},
-};
-pub(crate) mod request_block_bodies;
-use crate::{flow_context::FlowContext, flow_trait::Flow};
-
-use crate::ibd::IbdFlow;
-use kaspa_p2p_lib::{KaspadMessagePayloadType, Router, SharedIncomingRoute, convert::header::HeaderFormat};
-use kaspa_utils::channel;
-use request_block_bodies::HandleBlockBodyRequests;
-use std::sync::Arc;
-
-pub fn register(ctx: FlowContext, router: Arc<Router>, protocol_version: u32) -> Vec<Box<dyn Flow>> {
-    // IBD flow <-> invs flow communication uses a job channel in order to always
-    // maintain at most a single pending job which can be updated
-    let (ibd_sender, relay_receiver) = channel::job();
-    let body_only_ibd_permitted = true;
-    let header_format = HeaderFormat::from(protocol_version);
-    let mut flows: Vec<Box<dyn Flow>> = vec![
-        Box::new(IbdFlow::new(
-            ctx.clone(),
-            router.clone(),
-            router.subscribe(vec![
-                KaspadMessagePayloadType::BlockHeaders,
-                KaspadMessagePayloadType::DoneHeaders,
-                KaspadMessagePayloadType::IbdBlockLocatorHighestHash,
-                KaspadMessagePayloadType::IbdBlockLocatorHighestHashNotFound,
-                KaspadMessagePayloadType::BlockWithTrustedDataV4,
-                KaspadMessagePayloadType::DoneBlocksWithTrustedData,
-                KaspadMessagePayloadType::IbdChainBlockLocator,
-                KaspadMessagePayloadType::IbdBlock,
-                KaspadMessagePayloadType::BlockBody,
-                KaspadMessagePayloadType::TrustedData,
-                KaspadMessagePayloadType::PruningPoints,
-                KaspadMessagePayloadType::PruningPointProof,
-                KaspadMessagePayloadType::UnexpectedPruningPoint,
-                KaspadMessagePayloadType::PruningPointUtxoSetChunk,
-                KaspadMessagePayloadType::DonePruningPointUtxoSetChunks,
-            ]),
-            relay_receiver,
-            body_only_ibd_permitted,
-            header_format,
-        )),
-        Box::new(HandleRelayBlockRequests::new(
-            ctx.clone(),
-            router.clone(),
-            router.subscribe(vec![KaspadMessagePayloadType::RequestRelayBlocks]),
-            header_format,
-        )),
-        Box::new(ReceivePingsFlow::new(ctx.clone(), router.clone(), router.subscribe(vec![KaspadMessagePayloadType::Ping]))),
-        Box::new(SendPingsFlow::new(ctx.clone(), router.clone(), router.subscribe(vec![KaspadMessagePayloadType::Pong]))),
-        Box::new(RequestHeadersFlow::new(
-            ctx.clone(),
-            router.clone(),
-            router.subscribe(vec![KaspadMessagePayloadType::RequestHeaders, KaspadMessagePayloadType::RequestNextHeaders]),
-            header_format,
-        )),
-        Box::new(RequestPruningPointProofFlow::new(
-            ctx.clone(),
-            router.clone(),
-            router.subscribe(vec![KaspadMessagePayloadType::RequestPruningPointProof]),
-            header_format,
-        )),
-        Box::new(RequestIbdChainBlockLocatorFlow::new(
-            ctx.clone(),
-            router.clone(),
-            router.subscribe(vec![KaspadMessagePayloadType::RequestIbdChainBlockLocator]),
-        )),
-        Box::new(PruningPointAndItsAnticoneRequestsFlow::new(
-            ctx.clone(),
-            router.clone(),
-            router.subscribe(vec![
-                KaspadMessagePayloadType::RequestPruningPointAndItsAnticone,
-                KaspadMessagePayloadType::RequestNextPruningPointAndItsAnticoneBlocks,
-            ]),
-            header_format,
-        )),
-        Box::new(RequestPruningPointUtxoSetFlow::new(
-            ctx.clone(),
-            router.clone(),
-            router.subscribe(vec![
-                KaspadMessagePayloadType::RequestPruningPointUtxoSet,
-                KaspadMessagePayloadType::RequestNextPruningPointUtxoSetChunk,
-            ]),
-        )),
-        Box::new(HandleIbdBlockRequests::new(
-            ctx.clone(),
-            router.clone(),
-            router.subscribe(vec![KaspadMessagePayloadType::RequestIbdBlocks]),
-            header_format,
-        )),
-        Box::new(HandleBlockBodyRequests::new(
-            ctx.clone(),
-            router.clone(),
-            router.subscribe(vec![KaspadMessagePayloadType::RequestBlockBodies]),
-        )),
-        Box::new(HandleAntipastRequests::new(
-            ctx.clone(),
-            router.clone(),
-            router.subscribe(vec![KaspadMessagePayloadType::RequestAntipast]),
-            header_format,
-        )),
-        Box::new(RelayTransactionsFlow::new(
-            ctx.clone(),
-            router.clone(),
-            router
-                .subscribe_with_capacity(vec![KaspadMessagePayloadType::InvTransactions], RelayTransactionsFlow::invs_channel_size()),
-            router.subscribe_with_capacity(
-                vec![KaspadMessagePayloadType::Transaction, KaspadMessagePayloadType::TransactionNotFound],
-                RelayTransactionsFlow::txs_channel_size(),
-            ),
-        )),
-        Box::new(RequestTransactionsFlow::new(
-            ctx.clone(),
-            router.clone(),
-            router.subscribe(vec![KaspadMessagePayloadType::RequestTransactions]),
-        )),
-        Box::new(ReceiveAddressesFlow::new(ctx.clone(), router.clone(), router.subscribe(vec![KaspadMessagePayloadType::Addresses]))),
-        Box::new(SendAddressesFlow::new(
-            ctx.clone(),
-            router.clone(),
-            router.subscribe(vec![KaspadMessagePayloadType::RequestAddresses]),
-        )),
-        Box::new(RequestBlockLocatorFlow::new(
-            ctx.clone(),
-            router.clone(),
-            router.subscribe(vec![KaspadMessagePayloadType::RequestBlockLocator]),
-        )),
-    ];
-
-    let invs_route = router.subscribe_with_capacity(vec![KaspadMessagePayloadType::InvRelayBlock], ctx.block_invs_channel_size());
-    let shared_invs_route = SharedIncomingRoute::new(invs_route);
-
-    let num_relay_flows = (ctx.config.bps() as usize / 2).max(1);
-    flows.extend((0..num_relay_flows).map(|_| {
-        Box::new(HandleRelayInvsFlow::new(
-            ctx.clone(),
-            router.clone(),
-            shared_invs_route.clone(),
-            router.subscribe(vec![]),
-            ibd_sender.clone(),
-            header_format,
-        )) as Box<dyn Flow>
-    }));
-
-    // The reject message is handled as a special case by the router
-    // KaspadMessagePayloadType::Reject,
-
-    // We do not register the below two messages since they are deprecated also in go-kaspa
-    // KaspadMessagePayloadType::BlockWithTrustedData,
-    // KaspadMessagePayloadType::IbdBlockLocator,
-
-    flows
-}
